+++ CHANGELOG.md
@@ -1,5 +1,15 @@
 # Changelog
 
+## 4.1.5
+- Skinned mini meter: 
+- - always show player
+- - don't sometime stop when player die
+- - custom color bars when not class colored
+- - tooltips
+- - class colored names option
+- Added "Hide Pandemic Glow" toggle in CDM tab.
+- Moved CDM toggles to its own tab.
+
 ## 4.1.4
 - NB: Reworked pixels alignment and some sizing, so some things like icon groups size will be off and position to their anchor may be 1-2px off without an update\adjustment.
 - Fixed export string